Setting up for assembly dev
Gonna use armips to use ARM7TDMI Assembly (GBA Assembly) and FLoating IPS for patch generation I guess.

build.bat to build everything. Might build everything in a way to keep supporting both US and JP actually.

The 2 should be close enough for me to be able to do so, especially with all the free space I'd be able to get from removing all of 4Swords, who'd need that anyway.